+++ xgensemble.go
@@ -10,7 +10,7 @@ import (
 // XGEnsemble is XGBoost model (ensemble of trees)
 type XGEnsemble struct {
 	Trees         []lgTree
-	MaxFeatureIdx uint32
+	MaxFeatureIdx int
 }
 
 // NTrees returns number of trees in ensemble
@@ -24,7 +24,7 @@ func (e *XGEnsemble) NTrees() int {
 // Note, that result is a raw score (before sigmoid function transformation and etc).
 // Note, nan feature values treated as missing values
 func (e *XGEnsemble) Predict(fvals []float64, nTrees int) float64 {
-	if e.MaxFeatureIdx+1 > uint32(len(fvals)) {
+	if e.MaxFeatureIdx+1 > len(fvals) {
 		return 0.0
 	}
 	ret := 0.0
@@ -46,7 +46,7 @@ func (e *XGEnsemble) Predict(fvals []float64, nTrees int) float64 {
 // threads that will be utilized (maximum is GO_MAX_PROCS)
 // Note, that result is a raw score (before sigmoid function transformation and etc).
 // Note, `predictions` slice should be properly allocated on call side
-func (e *XGEnsemble) PredictCSR(indptr []uint32, cols []uint32, vals []float64, predictions []float64, nTrees int, nThreads int) {
+func (e *XGEnsemble) PredictCSR(indptr []int, cols []int, vals []float64, predictions []float64, nTrees int, nThreads int) {
 	nRows := len(indptr) - 1
 	if nRows <= BatchSize || nThreads == 0 || nThreads == 1 {
 		fvals := make([]float64, e.MaxFeatureIdx+1)
@@ -96,18 +96,18 @@ func (e *XGEnsemble) PredictCSR(indptr []uint32, cols []uint32, vals []float64, 
 	wg.Wait()
 }
 
-func (e *XGEnsemble) predictCSRInner(indptr []uint32, cols []uint32, vals []float64, startIndex int, endIndex int, predictions []float64, nTrees int, fvals []float64) {
+func (e *XGEnsemble) predictCSRInner(indptr []int, cols []int, vals []float64, startIndex int, endIndex int, predictions []float64, nTrees int, fvals []float64) {
 	for i := startIndex; i < endIndex; i++ {
 		start := indptr[i]
 		end := indptr[i+1]
 		for j := start; j < end; j++ {
-			if cols[j] < uint32(len(fvals)) {
+			if cols[j] < len(fvals) {
 				fvals[cols[j]] = vals[j]
 			}
 		}
 		predictions[i] = e.Predict(fvals, nTrees)
 		for j := start; j < end; j++ {
-			if cols[j] < uint32(len(fvals)) {
+			if cols[j] < len(fvals) {
 				fvals[cols[j]] = math.NaN()
 			}
 		}
@@ -120,8 +120,8 @@ func (e *XGEnsemble) predictCSRInner(indptr []uint32, cols []uint32, vals []floa
 // threads that will be utilized (maximum is GO_MAX_PROCS)
 // Note, that result is a raw score (before sigmoid function transformation and etc).
 // Note, `predictions` slice should be properly allocated on call side
-func (e *XGEnsemble) PredictDense(vals []float64, nrows uint32, ncols uint32, predictions []float64, nTrees int, nThreads int) error {
-	nRows := int(nrows)
+func (e *XGEnsemble) PredictDense(vals []float64, nrows int, ncols int, predictions []float64, nTrees int, nThreads int) error {
+	nRows := nrows
 	if ncols == 0 || e.MaxFeatureIdx > ncols-1 {
 		return fmt.Errorf("incorrect number of columns")
 	}
